% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regression_functions.R
\name{create_mvbf_formula}
\alias{create_mvbf_formula}
\title{Function to create a multivariate Bayesian formula for a given data frame.
This is an internal function.}
\usage{
create_mvbf_formula(df, group_vars = NULL, sigma = FALSE)
}
\arguments{
\item{df}{The data frame containing the estimates and SEs}

\item{group_vars}{Optional vector of grouping variables}

\item{sigma}{(Optional) Logical indicating whether to include sigma in the formula}
}
\value{
The multivariate formula for the brm function from brms package
}
\description{
Function to create a multivariate Bayesian formula for a given data frame.
This is an internal function.
}
