% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regression_functions.R
\name{extract_estimates}
\alias{extract_estimates}
\title{Function to extract estimates and standard errors from a list of fitted models}
\usage{
extract_estimates(fit_list, grouping_variable = NULL)
}
\arguments{
\item{fit_list}{A list of fitted models (e.g., ergm objects).}

\item{grouping_variable}{(Optional) A network-level attribute used for grouping.}
}
\value{
A tibble containing network id, coefficient id, coefficient term, estimate, and standard error.
}
\description{
Function to extract estimates and standard errors from a list of fitted models
}
