% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regression_functions.R
\name{meta_fit}
\alias{meta_fit}
\title{Function to fit a meta-analysis model using Bayesian regression}
\usage{
meta_fit(
  df,
  group_vars = NULL,
  chains = 4,
  cores = 4,
  iter = 4000,
  priors = prior("normal(0, 2)", class = "b"),
  backend = "cmdstanr",
  sigma = FALSE
)
}
\arguments{
\item{df}{A data frame containing the data for the meta-analysis}

\item{group_vars}{(optional) Variables indicating the grouping structure of the data}

\item{chains}{The number of Markov chains to run (default: 4)}

\item{cores}{The number of CPU cores to use for parallel computing (default: 4)}

\item{iter}{The number of iterations for each chain (default: 4000)}

\item{priors}{The prior distribution for the regression coefficients (default: normal(0, 2))}

\item{backend}{The backend for running the Bayesian regression (default: "cmdstanr")}

\item{sigma}{(Optional) Logical indicating whether to include sigma in the formula}
}
\value{
The fitted model object
}
\description{
Function to fit a meta-analysis model using Bayesian regression
}
