% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_estimation_functions.R
\name{ergm_estimation}
\alias{ergm_estimation}
\title{Function to estimate an Exponential Random Graph Model (ERGM) on a given network}
\usage{
ergm_estimation(network, formula, control_settings = control.ergm())
}
\arguments{
\item{network}{The input network object.}

\item{formula}{Formula specifying the ERGM model.}

\item{control_settings}{Optional control settings for the ERGM estimation. Default is control.ergm().}
}
\value{
The estimated ERGM model object or an informative error message.
}
\description{
This function takes a network object, an `ergm` formula, and optional control settings,
and estimates an ERGM model using the `ergm` function from the `ergm` package.
It handles potential convergence issues and returns the estimated model object or an informative error message.
}
