% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_estimation_functions.R
\name{ergm_estimation}
\alias{ergm_estimation}
\title{Function to estimate an Exponential Random Graph Model (ERGM) on a given network}
\usage{
ergm_estimation(network, formula, control_settings = NULL)
}
\arguments{
\item{network}{The input network object.}

\item{formula}{Formula for ERGM estimation.}

\item{contro_settings:}{Optional control settings for the ERGM estimation.}
}
\value{
The estimated ERGM model object.
}
\description{
Function to estimate an Exponential Random Graph Model (ERGM) on a given network
}
